requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
PyYAML>=6.0
pyarrow>=12.0.0
SQLAlchemy>=2.0.0
psycopg[binary]>=3.1.0
pydantic>=2.0.0
plotly>=5.17.0
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional but auto-detected
# xgboost>=2.0.0  # Uncomment if you want XGBoost support
# lightgbm>=4.0.0  # Uncomment if you want LightGBM support

# API server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Optional Supabase Storage
# supabase>=2.0.0  # Uncomment if you want Supabase Storage mirroring

# Optional FRED API
# requests>=2.31.0  # For FRED ingestion (included via other deps)

